# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CreateUserPushSettingRequest:
    properties:
      device_id:
        maxLength: 255
        minLength: 1
        type: string
      device_name:
        maxLength: 100
        minLength: 1
        type: string
      provider:
        enum:
        - bark
        type: string
      settings:
        additionalProperties: true
        type: object
    required:
    - device_id
    - device_name
    - provider
    type: object
  dto.ListResponse-dto_UserPushSettingResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.UserPushSettingResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  dto.UpdateUserPushSettingRequest:
    properties:
      device_name:
        type: string
      enabled:
        type: boolean
      settings:
        additionalProperties: true
        type: object
    type: object
...
  dto.UserPushSettingResponse:
    properties:
      created_at:
        type: string
      device_id:
        type: string
      device_name:
        type: string
      enabled:
        type: boolean
      id:
        type: integer
      provider:
        type: string
      settings:
        additionalProperties: true
        type: object
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  dto.ValidateDeviceRequest:
    properties:
      device_id:
        maxLength: 255
        minLength: 1
        type: string
      provider:
        enum:
        - bark
        type: string
    required:
    - device_id
    - provider
    type: object
  errors.APIError:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@nebula-live.com
    name: API Support
  description: Nebula Live is a modern backend API service built with Go, following
    Domain-Driven Design (DDD) principles and Clean Architecture.
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Nebula Live API
  version: "1.0"
paths:
...
  /push-settings:
    get:
      consumes:
      - application/json
      description: Get current user's push notification settings with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by provider
        enum:
        - bark
        in: query
        name: provider
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of user's push settings
          schema:
            $ref: '#/definitions/dto.ListResponse-dto_UserPushSettingResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - Bearer: []
      summary: Get Push Settings
      tags:
      - Push Settings
    post:
      consumes:
      - application/json
      description: Create a new push notification setting for current user
      parameters:
      - description: Push setting creation data
        in: body
        name: setting
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserPushSettingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Push setting created successfully
          schema:
            $ref: '#/definitions/dto.UserPushSettingResponse'
        "400":
          description: Invalid request parameters or validation failed
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "409":
          description: Device already exists
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - Bearer: []
      summary: Create Push Setting
      tags:
      - Push Settings
  /push-settings/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a push notification setting
      parameters:
      - description: Push setting ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Push setting deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid setting ID
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "404":
          description: Push setting not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - Bearer: []
      summary: Delete Push Setting
      tags:
      - Push Settings
    get:
      consumes:
      - application/json
      description: Get a specific push notification setting by ID
      parameters:
      - description: Push setting ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Push setting retrieved successfully
          schema:
            $ref: '#/definitions/dto.UserPushSettingResponse'
        "400":
          description: Invalid setting ID
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "404":
          description: Push setting not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - Bearer: []
      summary: Get Push Setting
      tags:
      - Push Settings
    put:
      consumes:
      - application/json
      description: Update a push notification setting
      parameters:
      - description: Push setting ID
        in: path
        name: id
        required: true
        type: integer
      - description: Push setting update data
        in: body
        name: setting
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserPushSettingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Push setting updated successfully
          schema:
            $ref: '#/definitions/dto.UserPushSettingResponse'
        "400":
          description: Invalid request parameters or validation failed
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "404":
          description: Push setting not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - Bearer: []
      summary: Update Push Setting
      tags:
      - Push Settings
  /push-settings/{id}/disable:
    post:
      consumes:
      - application/json
      description: Disable a push notification setting
      parameters:
      - description: Push setting ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Push setting disabled successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid setting ID
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "404":
          description: Push setting not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - Bearer: []
      summary: Disable Push Setting
      tags:
      - Push Settings
  /push-settings/{id}/enable:
    post:
      consumes:
      - application/json
      description: Enable a push notification setting
      parameters:
      - description: Push setting ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Push setting enabled successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid setting ID
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "404":
          description: Push setting not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - Bearer: []
      summary: Enable Push Setting
      tags:
      - Push Settings
...
  /push-settings/validate-device:
    post:
      consumes:
      - application/json
      description: Validate if a device ID is available for registration
      parameters:
      - description: Device validation data
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/dto.ValidateDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Device ID is available
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request parameters or validation failed
          schema:
            $ref: '#/definitions/errors.APIError'
        "409":
          description: Device ID already exists
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      summary: Validate Device ID
      tags:
      - Push Settings
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"